requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
faiss-cpu>=1.7.4
streamlit>=1.29.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
FlagEmbedding>=1.2.0
